+++ asr/sentence_recognizer.go
@@ -16,8 +16,8 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/Tencent-RTC/trtc-asr-sdk-go/common"
 	"github.com/google/uuid"
-	"github.com/hydah/trtc-asr-sdk-go/common"
 )
 
 // SentenceEndpoint is the production HTTPS endpoint for sentence recognition.