+++ examples/file_asr/main.go
@@ -24,8 +24,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/hydah/trtc-asr-sdk-go/asr"
-	"github.com/hydah/trtc-asr-sdk-go/common"
+	"github.com/Tencent-RTC/trtc-asr-sdk-go/asr"
+	"github.com/Tencent-RTC/trtc-asr-sdk-go/common"
 )
 
 // ===== Configuration =====