+++ examples/sentence_asr/main.go
@@ -19,8 +19,8 @@ import (
 	"log"
 	"os"
 
-	"github.com/hydah/trtc-asr-sdk-go/asr"
-	"github.com/hydah/trtc-asr-sdk-go/common"
+	"github.com/Tencent-RTC/trtc-asr-sdk-go/asr"
+	"github.com/Tencent-RTC/trtc-asr-sdk-go/common"
 )
 
 // ===== Configuration =====